feat: se Crearon los modulos de listarUsuarios, MostrarTareas en la carpeta Peticiones y se corrigio el controlador de casi todas las vistas menos vistaTarea en el enrutador
