+++ Cliente/Peticiones/ListarUsuario.js
@@ -2,6 +2,9 @@ import { request } from "../helper/index.js";
 
 export function listarUsuario(){
 
+    const userNameInput = document.querySelector('#userName');
+    const userNameList = document.getElementById('userNameList');
+    const UsuariosTableBody = document.querySelector('#UsuariosTableBody');
 
     function renderizarDatalist(users) {
         userNameList.innerHTML = '';
