+++ Cliente/VistaActualizar/ActualizarControlador.js
@@ -1,13 +1,16 @@
 import { actualizar } from "./mostrarActualizar.js"
 import { listarUsuario } from "../Peticiones/ListarUsuario.js"
+import { agregarTareas } from "../Peticiones/MostrarTareas.js";
 
 export const actualizarControlador = () => {
  
     const editable = document.querySelector('div > div');
     
-    const userNameList = document.getElementById('userNameList');
-    listarUsuario();
-
     editable.innerHTML = actualizar();
+    
+    
+    listarUsuario();
+    agregarTareas();
+    
     
 }
