+++ Cliente/VistaAgregar/AgregarControlador.js
@@ -1,12 +1,15 @@
 import { agregar } from "./mostrarAgregar.js"
 import { listarUsuario } from "../Peticiones/ListarUsuario.js"
+import { agregarTareas } from "../Peticiones/MostrarTareas.js";
 
 export const agregarControlador = () => {
  
     const editable = document.querySelector('div > div');
     const userNameList = document.getElementById('userNameList');
     
-    listarUsuario();
-    
     editable.innerHTML = agregar();
+    
+    listarUsuario();
+    agregarTareas();
+    
 }
