+++ Cliente/VistaEliminar/eliminarControlador.js
@@ -1,10 +1,13 @@
 import { eliminar } from "./mostrarEliminar.js"
 import { listarUsuario } from "../Peticiones/ListarUsuario.js"
+import { agregarTareas } from "../Peticiones/MostrarTareas.js";
 
 export const eliminarControlador = () => {
  
     const editable = document.querySelector('div > div');
-    listarUsuario();
-
     editable.innerHTML = eliminar();
+    
+    listarUsuario();
+    agregarTareas();
+
 }
