+++ Cliente/script.js
@@ -28,24 +28,24 @@ const eliminarTares = cambio.querySelector("#deleteForm")
 const actualizarTareas = cambio.querySelector("#updateForm")
 
 // Campos de entrada
-const userNameInput = cambio.querySelector('#userName');
-const userMessageInput = cambio.querySelector('#userMessage');
+// const userNameInput = cambio.querySelector('#userName');
+// const userMessageInput = cambio.querySelector('#userMessage');
 const userDesripcionInput = cambio.querySelector('#UserDescripcion');
-const conjuntoDatos = cambio.querySelector(".form__group");
+// const conjuntoDatos = cambio.querySelector(".form__group");
 
 const Deleteid = cambio.querySelector("#Deleteid");
-const aparecerDelete = cambio.querySelector("#aparecerDelete");
+// const aparecerDelete = cambio.querySelector("#aparecerDelete");
 
 const Updateid = cambio.querySelector("#updateid");
-const AparecerUpdate = cambio.querySelector("#aparecerUpdate");
+// const AparecerUpdate = cambio.querySelector("#aparecerUpdate");
 const ActualizarT = cambio.querySelector("#ActualizarT");
 
-// Botónes de envío
-const submitBtn = cambio.querySelector('#submitBtnid');
-const submitBtnTareas = cambio.querySelector('#submitBtn');
-const deleteBtn = cambio.querySelector("#deleteBtn");
-const updateBtn = cambio.querySelector("#updateBtn");
-const resetBtn = cambio.querySelector("#resetBtn")
+// // Botónes de envío
+// // const submitBtn = cambio.querySelector('#submitBtnid');
+// const submitBtnTareas = cambio.querySelector('#submitBtn');
+// const deleteBtn = cambio.querySelector("#deleteBtn");
+// const updateBtn = cambio.querySelector("#updateBtn");
+// const resetBtn = cambio.querySelector("#resetBtn")
 
 // Elementos para mostrar errores
 const userNameError = cambio.querySelector('#userNameError');
@@ -57,8 +57,8 @@ const idUpdateError = cambio.querySelector("#idUpdateError");
 // Contenedor donde se mostrarán los mensajes
 const messagesContainer = cambio.querySelector('#messagesContainer');
 
-// Estado vacío (mensaje que se muestra cuando no hay mensajes)
-const emptyState = cambio.querySelector('#emptyState');
+// // Estado vacío (mensaje que se muestra cuando no hay mensajes)
+// const emptyState = cambio.querySelector('#emptyState');
 
 // Contador de mensajes
 const messageCount = cambio.querySelector('#messageCount');
@@ -70,7 +70,7 @@ const messageCount = cambio.querySelector('#messageCount');
 
 
 // Variable para llevar el conteo de mensajes
-let totalMessages = 0;
+// let totalMessages = 0;
 
 window.addEventListener("hashchange", arrancar);
 
@@ -165,63 +165,63 @@ function comprobarEliminacion(a) {
 }
 
 
-// Funcion que valida que el usuario este registrado
-function validateId(a) {
+// // Funcion que valida que el usuario este registrado
+// function validateId(a) {
 
-    let isValid = true;
-    const userName = userNameInput.value;
+//     let isValid = true;
+//     const userName = userNameInput.value;
 
-    if (!isValidInput(userName)) {
-        userNameInput.classList.add("form__input--error");
-        userNameError.textContent = "El nombre de usuario no es valido"
-        isValid = false;
-    }
+//     if (!isValidInput(userName)) {
+//         userNameInput.classList.add("form__input--error");
+//         userNameError.textContent = "El nombre de usuario no es valido"
+//         isValid = false;
+//     }
 
-    else {
-        isValid = true
-        userNameInput.classList.remove("form__input--error");
-        userNameError.textContent = ""
-    }
+//     else {
+//         isValid = true
+//         userNameInput.classList.remove("form__input--error");
+//         userNameError.textContent = ""
+//     }
 
-    return isValid;
+//     return isValid;
 
-}
-
-
-
-// Agarra el tiempo actual
-function getCurrentTimestamp() {
-    const now = new Date();
-    const options = {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-        hour: '2-digit',
-        minute: '2-digit'
-    };
-    return now.toLocaleDateString('es-ES', options);
-}
+// }
 
 
 
-function updateMessageCount(totalMessages) {
-    // TODO: Implementar actualización del contador
-    // Pista: Usa template literals para crear el texto
-    // Formato: "X mensaje(s)" o "X mensajes"
+// // Agarra el tiempo actual
+// function getCurrentTimestamp() {
+//     const now = new Date();
+//     const options = {
+//         year: 'numeric',
+//         month: 'long',
+//         day: 'numeric',
+//         hour: '2-digit',
+//         minute: '2-digit'
+//     };
+//     return now.toLocaleDateString('es-ES', options);
+// }
 
 
-    messageCount.textContent = `${totalMessages} tarea/s`
-}
+
+// function updateMessageCount(totalMessages) {
+//     // TODO: Implementar actualización del contador
+//     // Pista: Usa template literals para crear el texto
+//     // Formato: "X mensaje(s)" o "X mensajes"
+
+
+//     messageCount.textContent = `${totalMessages} tarea/s`
+// }
 
 /**
  * Oculta el estado vacío (mensaje cuando no hay mensajes)
  */
-function hideEmptyState() {
-    // TODO: Implementar función para ocultar el estado vacío
-    // Pista: Agrega la clase 'hidden' al elemento emptyState
+// function hideEmptyState() {
+//     // TODO: Implementar función para ocultar el estado vacío
+//     // Pista: Agrega la clase 'hidden' al elemento emptyState
 
-    emptyState.classList.add("hidden")
-}
+//     emptyState.classList.add("hidden")
+// }
 
 /**
  * Muestra el estado vacío (mensaje cuando no hay mensajes)
@@ -299,47 +299,47 @@ function showEmptyState() {
 
 
 // funcion que me permite mostrar las tarjetas de los usuarios
-async function createMessageElement(userName) {
-    // TODO: Implementar la creación de un nuevo mensaje
+// async function createMessageElement(userName) {
+//     // TODO: Implementar la creación de un nuevo mensaje
 
-    messagesContainer.innerHTML = "";
-    // PASO 1: Crear el contenedor principal del mensaje
+//     messagesContainer.innerHTML = "";
+//     // PASO 1: Crear el contenedor principal del mensaje
 
-    const fecha = getCurrentTimestamp();
+//     const fecha = getCurrentTimestamp();
 
-    const datosUsuario = await obtenerTareasPorUsuario(userName);
-    // PASO 2: Crear la estructura HTML del mensaje
-    // Puedes usar innerHTML con la siguiente estructura:
+//     const datosUsuario = await obtenerTareasPorUsuario(userName);
+//     // PASO 2: Crear la estructura HTML del mensaje
+//     // Puedes usar innerHTML con la siguiente estructura:
 
-    datosUsuario.forEach(datos => {
-        const div = document.createElement("div");
+//     datosUsuario.forEach(datos => {
+//         const div = document.createElement("div");
 
-        // Asignar la clase 'message-card'
-        div.classList.add("message-card");
-        div.classList.add(`tarea${datos.id}`)
+//         // Asignar la clase 'message-card'
+//         div.classList.add("message-card");
+//         div.classList.add(`tarea${datos.id}`)
 
-        div.innerHTML = `
-            <div class="message-card__header">
-                <div class="message-card__user">
-                    <div class="message-card__avatar">${datos.id}</div>
-                    <span class="message-card__username">${datos.name}</span>
-                </div>
-                <span class="message-card__timestamp">${fecha}</span>
+//         div.innerHTML = `
+//             <div class="message-card__header">
+//                 <div class="message-card__user">
+//                     <div class="message-card__avatar">${datos.id}</div>
+//                     <span class="message-card__username">${datos.name}</span>
+//                 </div>
+//                 <span class="message-card__timestamp">${fecha}</span>
 
-                </div>
-                <div class="message-card__content">${datos.tarea}</div>
-                <div class="message-card__content">${datos.body}</div> 
-        `
+//                 </div>
+//                 <div class="message-card__content">${datos.tarea}</div>
+//                 <div class="message-card__content">${datos.body}</div> 
+//         `
 
-        messagesContainer.appendChild(div);
+//         messagesContainer.appendChild(div);
 
-        totalMessages += 1
+//         totalMessages += 1
 
-        updateMessageCount(totalMessages)
-    })
+//         updateMessageCount(totalMessages)
+//     })
 
-    hideEmptyState();
-}
+//     hideEmptyState();
+// }
 
 async function createTarjetas() {
 
@@ -385,56 +385,56 @@ async function createTarjetas() {
 // 4. MANEJO DE EVENTOS
 // ============================================
 
-async function handleFormSubmit() {
-    // TODO: Implementar el manejador del evento submit
+// async function handleFormSubmit() {
+//     // TODO: Implementar el manejador del evento submit
 
-    // PASO 2: Validar el formulario
-    const formularioV = validateId();
+//     // PASO 2: Validar el formulario
+//     const formularioV = validateId();
 
-    if (!formularioV) {
-        alert("Los datos ingresados no son validos")
-        return
-    }
+//     if (!formularioV) {
+//         alert("Los datos ingresados no son validos")
+//         return
+//     }
 
-    // PASO 3: Obtener los valores de los campos
-    const confirmacionUser = await BuscarNombre(userNameInput.value);
+//     // PASO 3: Obtener los valores de los campos
+//     const confirmacionUser = await BuscarNombre(userNameInput.value);
 
-    if (!confirmacionUser) {
-        alert("El usuario no existe")
-        return
-    }
-    else {
-        alert("el usuario existe");
-    }
+//     if (!confirmacionUser) {
+//         alert("El usuario no existe")
+//         return
+//     }
+//     else {
+//         alert("el usuario existe");
+//     }
 
-    //mostrar los campos habilitados
-    conjuntoDatos?.classList.add("form__id");
-    aparecerDelete?.classList.add("form__id");
-    AparecerUpdate?.classList.add("form__id");
+//     //mostrar los campos habilitados
+//     conjuntoDatos?.classList.add("form__id");
+//     aparecerDelete?.classList.add("form__id");
+//     AparecerUpdate?.classList.add("form__id");
     
-    //Mostar boton subir tareas
-    submitBtnTareas?.classList.remove("btn--secundary");
-    submitBtnTareas?.classList.add("btn--primary");
+//     //Mostar boton subir tareas
+//     submitBtnTareas?.classList.remove("btn--secundary");
+//     submitBtnTareas?.classList.add("btn--primary");
 
-    updateBtn?.classList.add("btn--primary");
-    updateBtn?.classList.remove("btn--secundary");
+//     updateBtn?.classList.add("btn--primary");
+//     updateBtn?.classList.remove("btn--secundary");
 
-    //Mostar boton eliminar tareas
-    deleteBtn?.classList.remove("btn--secundary");
-    deleteBtn?.classList.add("btn--primary");
+//     //Mostar boton eliminar tareas
+//     deleteBtn?.classList.remove("btn--secundary");
+//     deleteBtn?.classList.add("btn--primary");
 
-    //mostrar boton resetear formulario
-    resetBtn?.classList.remove("btn--secundary");
-    resetBtn?.classList.add("btn--primary");
+//     //mostrar boton resetear formulario
+//     resetBtn?.classList.remove("btn--secundary");
+//     resetBtn?.classList.add("btn--primary");
 
-    // Para inabilitar el campo nombre
-    userNameInput.setAttribute("disabled", "true");
-    console.log(userNameInput);
-    console.log(userMessageInput);
+//     // Para inabilitar el campo nombre
+//     userNameInput.setAttribute("disabled", "true");
+//     console.log(userNameInput);
+//     console.log(userMessageInput);
     
     
-    createMessageElement(userNameInput.value);
-}
+//     createMessageElement(userNameInput.value);
+// }
 
 async function AgregarTarjetas(event) {
 
@@ -543,14 +543,7 @@ async function eliminarTarea(event) {
 // 5. REGISTRO DE EVENTOS
 // ============================================
 
-cambio.addEventListener("click", (evento) => {
-    // Verificamos si el origen del evento fue nuestro input específico
-    if (evento.target && evento.target.id === "submitBtnid") {
 
-        evento.preventDefault();
-        handleFormSubmit();
-    }
-});
 
 cambio.addEventListener("submit", (evento) => {
     // Verificamos si el origen del evento fue nuestro input específico
